Added Spotify API calls, allowing playlists to be created when run.
When the script is run, if first asks the user to input a playlist name; this is a temporary feature until the previous HTML parser has been integrated. The user is then asked to input songs, which are sent over to Spotify through their API functions. This is repeated until the user submits the input 'stop'; this is also a temporary feature. Once done, the user's search results are added to the created playlist on Spotify.